+++ Code/Mantid/Framework/MDAlgorithms/src/CompositeImplicitFunctionParser.cpp
@@ -51,7 +51,7 @@ namespace Mantid
             boost::interprocess::unique_ptr<CompositeFunctionBuilder, Mantid::API::DeleterPolicy<CompositeFunctionBuilder> > functionBuilder(new CompositeFunctionBuilder);
             NodeList* childFunctionElementList = functionElement->childNodes();
            
-            for(size_t i = 0; i < childFunctionElementList->length(); i++)
+            for(unsigned long i = 0; i < childFunctionElementList->length(); i++)
             {
                 Element* childFunctionElement = dynamic_cast<Element*>(childFunctionElementList->item(i));
                 std::string typeName = childFunctionElement->localName();
